chore(codegen): regenerate pkg/ outputs (gofmt canonicalization + two encoders)
Regenerated with:
- resolveLayout fix (encoders with unbounded impls now pick newest layout)
- writeFile gofmt step (canonicalises all generated Go files)
- cz_req_takeoff_equip_all split into 0x0BAD + 0x0BF5 impls

Net behavioural changes (small, described in the semantics/codegen commits):
- pkg/encode/cz_req_guild_emblem_img2.go: [10]byte → [14]byte with the
  Unused field. Correct for pv >= 20190619 (goKore's target pv=20200401).
- pkg/encode/cz_req_takeoff_equip_all.go: [6]byte → []byte with pv-branching.
  Correct for both 0x0BAD (pv 20210818..20230905, 2 bytes) and 0x0BF5
  (pv >= 20230906, 6 bytes with Location). goKore doesn't use this packet
  at pv=20200401 (doesn't exist that pv), so the API change is safe.
- pkg/encode/register.go: adapt to the new []byte return of
  EncodeCzReqTakeoffEquipAll (was passing b[:], now returns directly).

Everything else in this diff is pure gofmt cosmetics (comment alignment,
const-block indentation).

+++ pkg/send/cz_req_random_upgrade_item.go
@@ -4,6 +4,6 @@ package send
 
 // CzReqRandomUpgradeItem is the request struct for the cz_req_random_upgrade_item action.
 type CzReqRandomUpgradeItem struct {
-	ItemId uint16
-	Index uint16
+	ItemId uint32
+	Index  uint16
 }
